+++ NEWS.md
@@ -1,3 +1,14 @@
+# voson.tcn 0.3.0
+
+## Major Changes
+- Added the `tcn_counts` function for retrieving tweet counts over time for conversation ids. This uses the API v2 `tweets/counts` endpoint and does not contribute to the monthly tweet cap. It could be used prior to collecting conversation threads to identify conversations to target or conversations to add to a skip list.
+
+## Minor Changes
+- Full-archive endpoint searches with the optional API v2 `max_results` querystring parameter have started to return response status code
+400 for values over 100.
+- The `max_results` parameter for function `tcn_threads` has been changed from the maximum total results, and now refers to and allows the API parameter of the same name above to be set. The default value has been set to 100. If left at default academic projects using the full-archive search endpoint will only collect 30,000 tweets per 15 minute rate-limit.
+- Previous `max_results` parameter for the function `tcn_threads` has been renamed to `max_total`.
+
 # voson.tcn 0.2.4
 
 ## Major Changes